requirements.txt
pytest==8.4.1
pytest-cov==6.2.1
numpy==2.3.2
scikit-learn==1.7.1
sentence-transformers==5.0.0
pandas==2.3.1